Delete the classic_version directory
and split test cases into separate files.
