+++ ut_frontend/l2tlb/ptw/test/bundle.py
@@ -1,0 +1,396 @@
+import toffee
+from toffee import Bundle, Signals, Signal
+
+class PTWBundle(Bundle):
+        
+    clock, \
+    reset, \
+    io_sfence_valid, \
+    io_csr_satp_mode, \
+    io_csr_satp_asid, \
+    io_csr_satp_ppn, \
+    io_csr_satp_changed, \
+    io_csr_vsatp_mode, \
+    io_csr_vsatp_asid, \
+    io_csr_vsatp_ppn, \
+    io_csr_vsatp_changed, \
+    io_csr_hgatp_mode, \
+    io_csr_hgatp_vmid, \
+    io_csr_hgatp_changed, \
+    io_csr_priv_mxr, \
+    io_csr_mPBMTE, \
+    io_csr_hPBMTE, \
+    io_req_ready, \
+    io_req_valid, \
+    io_req_bits_req_info_vpn, \
+    io_req_bits_req_info_s2xlate, \
+    io_req_bits_req_info_source, \
+    io_req_bits_l3Hit, \
+    io_req_bits_l2Hit, \
+    io_req_bits_ppn, \
+    io_req_bits_stage1Hit, \
+    io_req_bits_stage1_entry_0_tag, \
+    io_req_bits_stage1_entry_0_asid, \
+    io_req_bits_stage1_entry_0_vmid, \
+    io_req_bits_stage1_entry_0_n, \
+    io_req_bits_stage1_entry_0_pbmt, \
+    io_req_bits_stage1_entry_0_perm_d, \
+    io_req_bits_stage1_entry_0_perm_a, \
+    io_req_bits_stage1_entry_0_perm_g, \
+    io_req_bits_stage1_entry_0_perm_u, \
+    io_req_bits_stage1_entry_0_perm_x, \
+    io_req_bits_stage1_entry_0_perm_w, \
+    io_req_bits_stage1_entry_0_perm_r, \
+    io_req_bits_stage1_entry_0_level, \
+    io_req_bits_stage1_entry_0_v, \
+    io_req_bits_stage1_entry_0_ppn, \
+    io_req_bits_stage1_entry_0_ppn_low, \
+    io_req_bits_stage1_entry_0_pf, \
+    io_req_bits_stage1_entry_1_tag, \
+    io_req_bits_stage1_entry_1_asid, \
+    io_req_bits_stage1_entry_1_vmid, \
+    io_req_bits_stage1_entry_1_n, \
+    io_req_bits_stage1_entry_1_pbmt, \
+    io_req_bits_stage1_entry_1_perm_d, \
+    io_req_bits_stage1_entry_1_perm_a, \
+    io_req_bits_stage1_entry_1_perm_g, \
+    io_req_bits_stage1_entry_1_perm_u, \
+    io_req_bits_stage1_entry_1_perm_x, \
+    io_req_bits_stage1_entry_1_perm_w, \
+    io_req_bits_stage1_entry_1_perm_r, \
+    io_req_bits_stage1_entry_1_level, \
+    io_req_bits_stage1_entry_1_v, \
+    io_req_bits_stage1_entry_1_ppn, \
+    io_req_bits_stage1_entry_1_ppn_low, \
+    io_req_bits_stage1_entry_1_pf, \
+    io_req_bits_stage1_entry_2_tag, \
+    io_req_bits_stage1_entry_2_asid, \
+    io_req_bits_stage1_entry_2_vmid, \
+    io_req_bits_stage1_entry_2_n, \
+    io_req_bits_stage1_entry_2_pbmt, \
+    io_req_bits_stage1_entry_2_perm_d, \
+    io_req_bits_stage1_entry_2_perm_a, \
+    io_req_bits_stage1_entry_2_perm_g, \
+    io_req_bits_stage1_entry_2_perm_u, \
+    io_req_bits_stage1_entry_2_perm_x, \
+    io_req_bits_stage1_entry_2_perm_w, \
+    io_req_bits_stage1_entry_2_perm_r, \
+    io_req_bits_stage1_entry_2_level, \
+    io_req_bits_stage1_entry_2_v, \
+    io_req_bits_stage1_entry_2_ppn, \
+    io_req_bits_stage1_entry_2_ppn_low, \
+    io_req_bits_stage1_entry_2_pf, \
+    io_req_bits_stage1_entry_3_tag, \
+    io_req_bits_stage1_entry_3_asid, \
+    io_req_bits_stage1_entry_3_vmid, \
+    io_req_bits_stage1_entry_3_n, \
+    io_req_bits_stage1_entry_3_pbmt, \
+    io_req_bits_stage1_entry_3_perm_d, \
+    io_req_bits_stage1_entry_3_perm_a, \
+    io_req_bits_stage1_entry_3_perm_g, \
+    io_req_bits_stage1_entry_3_perm_u, \
+    io_req_bits_stage1_entry_3_perm_x, \
+    io_req_bits_stage1_entry_3_perm_w, \
+    io_req_bits_stage1_entry_3_perm_r, \
+    io_req_bits_stage1_entry_3_level, \
+    io_req_bits_stage1_entry_3_v, \
+    io_req_bits_stage1_entry_3_ppn, \
+    io_req_bits_stage1_entry_3_ppn_low, \
+    io_req_bits_stage1_entry_3_pf, \
+    io_req_bits_stage1_entry_4_tag, \
+    io_req_bits_stage1_entry_4_asid, \
+    io_req_bits_stage1_entry_4_vmid, \
+    io_req_bits_stage1_entry_4_n, \
+    io_req_bits_stage1_entry_4_pbmt, \
+    io_req_bits_stage1_entry_4_perm_d, \
+    io_req_bits_stage1_entry_4_perm_a, \
+    io_req_bits_stage1_entry_4_perm_g, \
+    io_req_bits_stage1_entry_4_perm_u, \
+    io_req_bits_stage1_entry_4_perm_x, \
+    io_req_bits_stage1_entry_4_perm_w, \
+    io_req_bits_stage1_entry_4_perm_r, \
+    io_req_bits_stage1_entry_4_level, \
+    io_req_bits_stage1_entry_4_v, \
+    io_req_bits_stage1_entry_4_ppn, \
+    io_req_bits_stage1_entry_4_ppn_low, \
+    io_req_bits_stage1_entry_4_pf, \
+    io_req_bits_stage1_entry_5_tag, \
+    io_req_bits_stage1_entry_5_asid, \
+    io_req_bits_stage1_entry_5_vmid, \
+    io_req_bits_stage1_entry_5_n, \
+    io_req_bits_stage1_entry_5_pbmt, \
+    io_req_bits_stage1_entry_5_perm_d, \
+    io_req_bits_stage1_entry_5_perm_a, \
+    io_req_bits_stage1_entry_5_perm_g, \
+    io_req_bits_stage1_entry_5_perm_u, \
+    io_req_bits_stage1_entry_5_perm_x, \
+    io_req_bits_stage1_entry_5_perm_w, \
+    io_req_bits_stage1_entry_5_perm_r, \
+    io_req_bits_stage1_entry_5_level, \
+    io_req_bits_stage1_entry_5_v, \
+    io_req_bits_stage1_entry_5_ppn, \
+    io_req_bits_stage1_entry_5_ppn_low, \
+    io_req_bits_stage1_entry_5_pf, \
+    io_req_bits_stage1_entry_6_tag, \
+    io_req_bits_stage1_entry_6_asid, \
+    io_req_bits_stage1_entry_6_vmid, \
+    io_req_bits_stage1_entry_6_n, \
+    io_req_bits_stage1_entry_6_pbmt, \
+    io_req_bits_stage1_entry_6_perm_d, \
+    io_req_bits_stage1_entry_6_perm_a, \
+    io_req_bits_stage1_entry_6_perm_g, \
+    io_req_bits_stage1_entry_6_perm_u, \
+    io_req_bits_stage1_entry_6_perm_x, \
+    io_req_bits_stage1_entry_6_perm_w, \
+    io_req_bits_stage1_entry_6_perm_r, \
+    io_req_bits_stage1_entry_6_level, \
+    io_req_bits_stage1_entry_6_v, \
+    io_req_bits_stage1_entry_6_ppn, \
+    io_req_bits_stage1_entry_6_ppn_low, \
+    io_req_bits_stage1_entry_6_pf, \
+    io_req_bits_stage1_entry_7_tag, \
+    io_req_bits_stage1_entry_7_asid, \
+    io_req_bits_stage1_entry_7_vmid, \
+    io_req_bits_stage1_entry_7_n, \
+    io_req_bits_stage1_entry_7_pbmt, \
+    io_req_bits_stage1_entry_7_perm_d, \
+    io_req_bits_stage1_entry_7_perm_a, \
+    io_req_bits_stage1_entry_7_perm_g, \
+    io_req_bits_stage1_entry_7_perm_u, \
+    io_req_bits_stage1_entry_7_perm_x, \
+    io_req_bits_stage1_entry_7_perm_w, \
+    io_req_bits_stage1_entry_7_perm_r, \
+    io_req_bits_stage1_entry_7_level, \
+    io_req_bits_stage1_entry_7_v, \
+    io_req_bits_stage1_entry_7_ppn, \
+    io_req_bits_stage1_entry_7_ppn_low, \
+    io_req_bits_stage1_entry_7_pf, \
+    io_req_bits_stage1_pteidx_0, \
+    io_req_bits_stage1_pteidx_1, \
+    io_req_bits_stage1_pteidx_2, \
+    io_req_bits_stage1_pteidx_3, \
+    io_req_bits_stage1_pteidx_4, \
+    io_req_bits_stage1_pteidx_5, \
+    io_req_bits_stage1_pteidx_6, \
+    io_req_bits_stage1_pteidx_7, \
+    io_req_bits_stage1_not_super, \
+    io_resp_ready, \
+    io_resp_valid, \
+    io_resp_bits_source, \
+    io_resp_bits_s2xlate, \
+    io_resp_bits_resp_entry_0_tag, \
+    io_resp_bits_resp_entry_0_asid, \
+    io_resp_bits_resp_entry_0_vmid, \
+    io_resp_bits_resp_entry_0_n, \
+    io_resp_bits_resp_entry_0_pbmt, \
+    io_resp_bits_resp_entry_0_perm_d, \
+    io_resp_bits_resp_entry_0_perm_a, \
+    io_resp_bits_resp_entry_0_perm_g, \
+    io_resp_bits_resp_entry_0_perm_u, \
+    io_resp_bits_resp_entry_0_perm_x, \
+    io_resp_bits_resp_entry_0_perm_w, \
+    io_resp_bits_resp_entry_0_perm_r, \
+    io_resp_bits_resp_entry_0_level, \
+    io_resp_bits_resp_entry_0_v, \
+    io_resp_bits_resp_entry_0_ppn, \
+    io_resp_bits_resp_entry_0_ppn_low, \
+    io_resp_bits_resp_entry_0_af, \
+    io_resp_bits_resp_entry_0_pf, \
+    io_resp_bits_resp_entry_1_tag, \
+    io_resp_bits_resp_entry_1_asid, \
+    io_resp_bits_resp_entry_1_vmid, \
+    io_resp_bits_resp_entry_1_n, \
+    io_resp_bits_resp_entry_1_pbmt, \
+    io_resp_bits_resp_entry_1_perm_d, \
+    io_resp_bits_resp_entry_1_perm_a, \
+    io_resp_bits_resp_entry_1_perm_g, \
+    io_resp_bits_resp_entry_1_perm_u, \
+    io_resp_bits_resp_entry_1_perm_x, \
+    io_resp_bits_resp_entry_1_perm_w, \
+    io_resp_bits_resp_entry_1_perm_r, \
+    io_resp_bits_resp_entry_1_level, \
+    io_resp_bits_resp_entry_1_v, \
+    io_resp_bits_resp_entry_1_ppn, \
+    io_resp_bits_resp_entry_1_ppn_low, \
+    io_resp_bits_resp_entry_1_af, \
+    io_resp_bits_resp_entry_1_pf, \
+    io_resp_bits_resp_entry_2_tag, \
+    io_resp_bits_resp_entry_2_asid, \
+    io_resp_bits_resp_entry_2_vmid, \
+    io_resp_bits_resp_entry_2_n, \
+    io_resp_bits_resp_entry_2_pbmt, \
+    io_resp_bits_resp_entry_2_perm_d, \
+    io_resp_bits_resp_entry_2_perm_a, \
+    io_resp_bits_resp_entry_2_perm_g, \
+    io_resp_bits_resp_entry_2_perm_u, \
+    io_resp_bits_resp_entry_2_perm_x, \
+    io_resp_bits_resp_entry_2_perm_w, \
+    io_resp_bits_resp_entry_2_perm_r, \
+    io_resp_bits_resp_entry_2_level, \
+    io_resp_bits_resp_entry_2_v, \
+    io_resp_bits_resp_entry_2_ppn, \
+    io_resp_bits_resp_entry_2_ppn_low, \
+    io_resp_bits_resp_entry_2_af, \
+    io_resp_bits_resp_entry_2_pf, \
+    io_resp_bits_resp_entry_3_tag, \
+    io_resp_bits_resp_entry_3_asid, \
+    io_resp_bits_resp_entry_3_vmid, \
+    io_resp_bits_resp_entry_3_n, \
+    io_resp_bits_resp_entry_3_pbmt, \
+    io_resp_bits_resp_entry_3_perm_d, \
+    io_resp_bits_resp_entry_3_perm_a, \
+    io_resp_bits_resp_entry_3_perm_g, \
+    io_resp_bits_resp_entry_3_perm_u, \
+    io_resp_bits_resp_entry_3_perm_x, \
+    io_resp_bits_resp_entry_3_perm_w, \
+    io_resp_bits_resp_entry_3_perm_r, \
+    io_resp_bits_resp_entry_3_level, \
+    io_resp_bits_resp_entry_3_v, \
+    io_resp_bits_resp_entry_3_ppn, \
+    io_resp_bits_resp_entry_3_ppn_low, \
+    io_resp_bits_resp_entry_3_af, \
+    io_resp_bits_resp_entry_3_pf, \
+    io_resp_bits_resp_entry_4_tag, \
+    io_resp_bits_resp_entry_4_asid, \
+    io_resp_bits_resp_entry_4_vmid, \
+    io_resp_bits_resp_entry_4_n, \
+    io_resp_bits_resp_entry_4_pbmt, \
+    io_resp_bits_resp_entry_4_perm_d, \
+    io_resp_bits_resp_entry_4_perm_a, \
+    io_resp_bits_resp_entry_4_perm_g, \
+    io_resp_bits_resp_entry_4_perm_u, \
+    io_resp_bits_resp_entry_4_perm_x, \
+    io_resp_bits_resp_entry_4_perm_w, \
+    io_resp_bits_resp_entry_4_perm_r, \
+    io_resp_bits_resp_entry_4_level, \
+    io_resp_bits_resp_entry_4_v, \
+    io_resp_bits_resp_entry_4_ppn, \
+    io_resp_bits_resp_entry_4_ppn_low, \
+    io_resp_bits_resp_entry_4_af, \
+    io_resp_bits_resp_entry_4_pf, \
+    io_resp_bits_resp_entry_5_tag, \
+    io_resp_bits_resp_entry_5_asid, \
+    io_resp_bits_resp_entry_5_vmid, \
+    io_resp_bits_resp_entry_5_n, \
+    io_resp_bits_resp_entry_5_pbmt, \
+    io_resp_bits_resp_entry_5_perm_d, \
+    io_resp_bits_resp_entry_5_perm_a, \
+    io_resp_bits_resp_entry_5_perm_g, \
+    io_resp_bits_resp_entry_5_perm_u, \
+    io_resp_bits_resp_entry_5_perm_x, \
+    io_resp_bits_resp_entry_5_perm_w, \
+    io_resp_bits_resp_entry_5_perm_r, \
+    io_resp_bits_resp_entry_5_level, \
+    io_resp_bits_resp_entry_5_v, \
+    io_resp_bits_resp_entry_5_ppn, \
+    io_resp_bits_resp_entry_5_ppn_low, \
+    io_resp_bits_resp_entry_5_af, \
+    io_resp_bits_resp_entry_5_pf, \
+    io_resp_bits_resp_entry_6_tag, \
+    io_resp_bits_resp_entry_6_asid, \
+    io_resp_bits_resp_entry_6_vmid, \
+    io_resp_bits_resp_entry_6_n, \
+    io_resp_bits_resp_entry_6_pbmt, \
+    io_resp_bits_resp_entry_6_perm_d, \
+    io_resp_bits_resp_entry_6_perm_a, \
+    io_resp_bits_resp_entry_6_perm_g, \
+    io_resp_bits_resp_entry_6_perm_u, \
+    io_resp_bits_resp_entry_6_perm_x, \
+    io_resp_bits_resp_entry_6_perm_w, \
+    io_resp_bits_resp_entry_6_perm_r, \
+    io_resp_bits_resp_entry_6_level, \
+    io_resp_bits_resp_entry_6_v, \
+    io_resp_bits_resp_entry_6_ppn, \
+    io_resp_bits_resp_entry_6_ppn_low, \
+    io_resp_bits_resp_entry_6_af, \
+    io_resp_bits_resp_entry_6_pf, \
+    io_resp_bits_resp_entry_7_tag, \
+    io_resp_bits_resp_entry_7_asid, \
+    io_resp_bits_resp_entry_7_vmid, \
+    io_resp_bits_resp_entry_7_n, \
+    io_resp_bits_resp_entry_7_pbmt, \
+    io_resp_bits_resp_entry_7_perm_d, \
+    io_resp_bits_resp_entry_7_perm_a, \
+    io_resp_bits_resp_entry_7_perm_g, \
+    io_resp_bits_resp_entry_7_perm_u, \
+    io_resp_bits_resp_entry_7_perm_x, \
+    io_resp_bits_resp_entry_7_perm_w, \
+    io_resp_bits_resp_entry_7_perm_r, \
+    io_resp_bits_resp_entry_7_level, \
+    io_resp_bits_resp_entry_7_v, \
+    io_resp_bits_resp_entry_7_ppn, \
+    io_resp_bits_resp_entry_7_ppn_low, \
+    io_resp_bits_resp_entry_7_af, \
+    io_resp_bits_resp_entry_7_pf, \
+    io_resp_bits_resp_pteidx_0, \
+    io_resp_bits_resp_pteidx_1, \
+    io_resp_bits_resp_pteidx_2, \
+    io_resp_bits_resp_pteidx_3, \
+    io_resp_bits_resp_pteidx_4, \
+    io_resp_bits_resp_pteidx_5, \
+    io_resp_bits_resp_pteidx_6, \
+    io_resp_bits_resp_pteidx_7, \
+    io_resp_bits_resp_not_super, \
+    io_resp_bits_h_resp_entry_tag, \
+    io_resp_bits_h_resp_entry_vmid, \
+    io_resp_bits_h_resp_entry_n, \
+    io_resp_bits_h_resp_entry_pbmt, \
+    io_resp_bits_h_resp_entry_ppn, \
+    io_resp_bits_h_resp_entry_perm_d, \
+    io_resp_bits_h_resp_entry_perm_a, \
+    io_resp_bits_h_resp_entry_perm_g, \
+    io_resp_bits_h_resp_entry_perm_u, \
+    io_resp_bits_h_resp_entry_perm_x, \
+    io_resp_bits_h_resp_entry_perm_w, \
+    io_resp_bits_h_resp_entry_perm_r, \
+    io_resp_bits_h_resp_entry_level, \
+    io_resp_bits_h_resp_gpf, \
+    io_resp_bits_h_resp_gaf, \
+    io_llptw_ready, \
+    io_llptw_valid, \
+    io_llptw_bits_req_info_vpn, \
+    io_llptw_bits_req_info_s2xlate, \
+    io_llptw_bits_req_info_source, \
+    io_hptw_req_ready, \
+    io_hptw_req_valid, \
+    io_hptw_req_bits_source, \
+    io_hptw_req_bits_gvpn, \
+    io_hptw_resp_valid, \
+    io_hptw_resp_bits_h_resp_entry_tag, \
+    io_hptw_resp_bits_h_resp_entry_vmid, \
+    io_hptw_resp_bits_h_resp_entry_n, \
+    io_hptw_resp_bits_h_resp_entry_pbmt, \
+    io_hptw_resp_bits_h_resp_entry_ppn, \
+    io_hptw_resp_bits_h_resp_entry_perm_d, \
+    io_hptw_resp_bits_h_resp_entry_perm_a, \
+    io_hptw_resp_bits_h_resp_entry_perm_g, \
+    io_hptw_resp_bits_h_resp_entry_perm_u, \
+    io_hptw_resp_bits_h_resp_entry_perm_x, \
+    io_hptw_resp_bits_h_resp_entry_perm_w, \
+    io_hptw_resp_bits_h_resp_entry_perm_r, \
+    io_hptw_resp_bits_h_resp_entry_level, \
+    io_hptw_resp_bits_h_resp_gpf, \
+    io_hptw_resp_bits_h_resp_gaf, \
+    io_mem_req_ready, \
+    io_mem_req_valid, \
+    io_mem_req_bits_addr, \
+    io_mem_resp_valid, \
+    io_mem_resp_bits, \
+    io_mem_mask, \
+    io_pmp_req_bits_addr, \
+    io_pmp_resp_ld, \
+    io_pmp_resp_mmio, \
+    io_refill_req_info_vpn, \
+    io_refill_req_info_s2xlate, \
+    io_refill_req_info_source, \
+    io_refill_level, \
+    io_perf_0_value, \
+    io_perf_1_value, \
+    io_perf_2_value, \
+    io_perf_3_value, \
+    io_perf_4_value, \
+    io_perf_5_value, \
+    io_perf_6_value = Signals(388)
+  
+
+
